code refactoring
Move all the endpoint to lib.rs in this way they can be easly tested.
